Update 2020-06-18-rose-complex-types.markdown
Add explanation as to why the source location is output as 0 for the translator in Section C.

I think it would be a good idea to have an exercise as part of the practicals, once available, for making the translator code more robust. It would help reinforce what has been learned.