Add practical for complex types tutorial
Content is laid out and one of two components of the practical is completed: users can build their tool with the supplies code and makefile.

Next, we need to add check rules to the makefile and compare the output with diff to the expected output to report on the user's correctness.