Update contribute.md
Using a place holder of the language such as <language> messes up jekyll serve and wont allow it to restart (it seems jekyll thinks that we're trying to insert real code and it doesn't recognize <language> as a supported lanaguge).

Instead, rewrite to avoid this by having the user replace cpp with language of their choice.