Add code snippets to tutorials with collapse support
As per Dr. Liao's feedback, add code snippets in the tutorial text whenever we use cat in the terminal. This ensures that users do not *have* to rely on the terminal, but could also just read the text.

To keep the tutorial text from growing too large, make these some of these code snippets collapsible, and closed by default. Users can just click to open to view the source code. This is the same number of clicks as when using the terminal (to click the cat command).

Restructure some tutorial texts to support this feature.