Add code container for those using explicit line numbers
Jekyll unfortunately does not have support for starting line numbers at a certain number. So we have to use normal code containers and just manually type the lines.

Using the three back ticks makes this ugly because the container is fluid. Instead, this CSS prevents that.

Code using this container will have to be placed within liquid tags, and a surrounding <figure> element will have to be placed. The <figure> element must use this class. The CSS also styles the container to be consistent with the others.

Overall, this improves readibility.