Revise upload request
Update endpoint, support file path, use ajax instead of XMLHttpRequest.

The problem with using XMLHttpRequest is that it will not let us define an undefined content-type. Doing this lets the browser automatically determine the type. The upload endpoint accepts multipart content-type, but I was unable to get the boundary param to work, so the easiest solution is to let the browser do it.

A workaround is to just use an ajax call and specify content-type: false to have the browser setup the multipart for us. Maybe there is a way to do the same with XMLHttpRequest, or we could see if we can switch the other functions to use ajax.