Add layout for closed-book testing posts
This is the client side component of closed-book testing: it communicates with the PWC backend to send the user's code to be compiled.

Some of it will need to be extended once the backend auto-grading is completed, depending on how we handle that request wise.

Relevant CSS included as well.