Support downloading files directly from sandbox client-side
References https://github.com/chunhualiao/freeCompilerCamp/issues/38.

See discussion there. This approach is flexible enough to support downloading any type of files that have a supported MIME type. It is easy to create as a contributor.

Add a link to download a generated PDF in the AST intro tutorial, directly from the sandbox, in this commit.